# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeanAI Assistance API",
        "description": "Speech practice and AI-powered analysis API for SpeechFlow. Provides authentication, challenge browsing, practice session management, and AI-driven speech recording analysis.",
        "contact": {
            "name": "API Support",
            "email": "support@speechflow.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/challenges": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "List challenges",
                "description": "Returns challenges with optional category and level filtering.",
                "operationId": "5f7157215899fabbbf767e5b914a5f47",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "public-speaking"
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by level slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "beginner"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of challenges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Challenge"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category or level not found"
                    }
                }
            }
        },
        "/challenges/{id}": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Get a single challenge",
                "description": "Returns a specific challenge by ID, including its category and level.",
                "operationId": "909afc90fe083c4d9022d73558a4a4d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Challenge ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Challenge details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Challenge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge not found"
                    }
                }
            }
        },
        "/practice-sessions": {
            "get": {
                "tags": [
                    "Practice Sessions"
                ],
                "summary": "List practice sessions",
                "description": "Returns all practice sessions for the authenticated user, ordered by most recent first.",
                "operationId": "aabedfb320883069d010174a4062fcf4",
                "responses": {
                    "200": {
                        "description": "List of sessions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PracticeSession"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Practice Sessions"
                ],
                "summary": "Start a new practice session",
                "description": "Creates a new practice session for the authenticated user, linked to a specific challenge.",
                "operationId": "116f4435e0236f25425d559cdb8f3876",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "challenge_id"
                                ],
                                "properties": {
                                    "challenge_id": {
                                        "description": "ID of the challenge to practice",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Optional custom session name. Auto-generated if omitted.",
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "My morning practice"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PracticeSession"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error (invalid challenge_id)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/practice-sessions/{id}": {
            "get": {
                "tags": [
                    "Practice Sessions"
                ],
                "summary": "Get a practice session",
                "description": "Returns a specific practice session with its challenge and result.",
                "operationId": "df1fc017e68150a2d33480874c0746ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Practice session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PracticeSession"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "public-speaking"
                    },
                    "name": {
                        "type": "string",
                        "example": "Public Speaking"
                    },
                    "description": {
                        "type": "string",
                        "example": "Practice your public speaking skills"
                    },
                    "icon": {
                        "type": "string",
                        "example": "microphone"
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Challenge": {
                "title": "Challenge",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduce yourself to a new team"
                    },
                    "suggested_time_minutes": {
                        "type": "integer",
                        "example": 5
                    },
                    "hints_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "tips": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Speak clearly",
                            "Make eye contact"
                        ]
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            }
                        ],
                        "nullable": true
                    },
                    "level": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Level"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Level": {
                "title": "Level",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "beginner"
                    },
                    "name": {
                        "type": "string",
                        "example": "Beginner"
                    },
                    "description": {
                        "type": "string",
                        "example": "For those just starting out"
                    },
                    "color": {
                        "type": "string",
                        "example": "#4CAF50"
                    },
                    "min_score": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PracticeSession": {
                "title": "PracticeSession",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Introduce yourself Practice - Feb 08, 2026 10:30 AM"
                    },
                    "challenge": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Challenge"
                            }
                        ],
                        "nullable": true
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "started",
                            "recorded",
                            "processed",
                            "failed"
                        ],
                        "example": "processed"
                    },
                    "result": {
                        "description": "Included when the result relationship is loaded",
                        "properties": {
                            "transcription": {
                                "type": "string"
                            },
                            "feedback": {
                                "type": "string"
                            },
                            "score": {
                                "type": "number",
                                "format": "float"
                            },
                            "metadata": {
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Bearer token obtained from /auth/login or /auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}